# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LedgerLoop API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/ai-categories/accept-single": {
      "post": {
        "tags": [
          "ai-categories"
        ],
        "summary": "Accept Single Category",
        "description": "Accept a single AI category suggestion with smart duplicate handling",
        "operationId": "accept_single_category_api_ai_categories_accept_single_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SingleAcceptanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-categories/accept-batch": {
      "post": {
        "tags": [
          "ai-categories"
        ],
        "summary": "Accept Batch Categories",
        "description": "Accept multiple AI category suggestions in batch - 'Accept All' functionality",
        "operationId": "accept_batch_categories_api_ai_categories_accept_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAcceptanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ai-categories/preview-batch": {
      "post": {
        "tags": [
          "ai-categories"
        ],
        "summary": "Preview Batch Acceptance",
        "description": "Preview what would happen with batch acceptance without actually doing it",
        "operationId": "preview_batch_acceptance_api_ai_categories_preview_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAcceptanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-categories/categorize-transactions": {
      "post": {
        "tags": [
          "ai-categories"
        ],
        "summary": "Categorize Transactions",
        "description": "Get AI categorization suggestions for multiple transactions",
        "operationId": "categorize_transactions_api_ai_categories_categorize_transactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCategorizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-categories/suggestions-status/{transaction_id}": {
      "get": {
        "tags": [
          "ai-categories"
        ],
        "summary": "Get Suggestions Status",
        "description": "Get the status of AI suggestions for a specific transaction",
        "operationId": "get_suggestions_status_api_ai_categories_suggestions_status__transaction_id__get",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-categories/auto-categorize-uncategorized": {
      "post": {
        "tags": [
          "ai-categories"
        ],
        "summary": "Auto Categorize Uncategorized",
        "description": "Auto-categorize uncategorized transactions with high confidence",
        "operationId": "auto_categorize_uncategorized_api_ai_categories_auto_categorize_uncategorized_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "confidence_threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.85,
              "title": "Confidence Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchAcceptanceRequest": {
        "properties": {
          "acceptances": {
            "items": {
              "$ref": "#/components/schemas/SingleAcceptanceRequest"
            },
            "type": "array",
            "title": "Acceptances"
          },
          "create_missing_categories": {
            "type": "boolean",
            "title": "Create Missing Categories",
            "default": true
          },
          "auto_merge_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Auto Merge Threshold",
            "default": 0.85
          },
          "create_parent_hierarchy": {
            "type": "boolean",
            "title": "Create Parent Hierarchy",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "acceptances"
        ],
        "title": "BatchAcceptanceRequest",
        "description": "Request to accept multiple AI category suggestions."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SingleAcceptanceRequest": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "suggested_category_name": {
            "type": "string",
            "title": "Suggested Category Name"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "create_if_missing": {
            "type": "boolean",
            "title": "Create If Missing",
            "default": true
          },
          "auto_merge_similar": {
            "type": "boolean",
            "title": "Auto Merge Similar",
            "default": true
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "suggested_category_name",
          "confidence"
        ],
        "title": "SingleAcceptanceRequest",
        "description": "Request to accept a single AI category suggestion."
      },
...
      "TransactionCategorizeRequest": {
        "properties": {
          "transaction_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Transaction Ids"
          },
          "apply_high_confidence": {
            "type": "boolean",
            "title": "Apply High Confidence",
            "default": false
          },
          "confidence_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Threshold",
            "default": 0.85
          }
        },
        "type": "object",
        "required": [
          "transaction_ids"
        ],
        "title": "TransactionCategorizeRequest",
        "description": "Request to get AI categorization for transactions"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}